Refactor Geohashing class to geohashing.py
